Removed config.rs and put back actix config in main.rs
